move objectToError to utils
